Add Trino experimental SPIs usage checker
Bug checker for detecting usages of Trino SPIs that are annotated with
the `@Experimental` annotation. Such SPIs should be avoided in
libraries and shared components that other projects depend on. This
checker is intended to be used by such libraries and shared components
to ensure that they don't use them.