Only check ignored types for symbols representing types
When we check the set of ignored types, we do it for all encountered
symbols. A conflict is unlikely, because the type names are fully
qualified, and e.g. fields and methods are not, but I still feel that
it's better to only do it for actual types.

With the compiler API it's enough to check that the kind is `CLASS`, but
tests are enhanced to include also `enum` and `record` types just to be
sure.