Upgrade Error Prone to 2.19.1
The constructor resolution logic got changed in Error Prone 2.19.1 and
now we have to explicitly designate the constructor that should be used.
The issue when there's no `@Inject` annotation is that it will look for
the constructor which has all parameters of type `ErrorProneFlags` - but
the `allMatch` operator will also return `true` when the list of
parameters is empty, so it may also pick up the no-args constructor.